Update CCF to use a more traditional CCF build process
Updated the build process for the PDO TP so that it
should not require any modification to the CCF source.
This is the normal way for building CCF applications.

Updated most of the scripts to simplify starting and
stopping a service. The service will now execute out
of an installed directory $PDO_HOME/ccf/*.

Cleaned up most of the scripts to use the configuration
file provided by cchost in order to find end point
information.

Placed all keys in one canoncical directory. I do not
know if there is any danger in exposing the member
private key. We need to evaluate this.

Signed-off-by: Mic Bowman <cmickeyb@gmail.com>